pyproject.toml
[project]
name = "compare-python-processing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.0",
    "numba>=0.60",
    "tabulate>=0.9",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "maturin>=1.12.6",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.15.7",
    "types-pyyaml>=6.0.12.20250915",
    "types-tabulate>=0.10.0.20260308",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
strict = true
python_version = "3.12"

[tool.pytest.ini_options]
testpaths = ["tests"]
